minor tidy
