+++ llms/providers/__init__.py
@@ -14,3 +14,22 @@ from .openai import OpenAIProvider
 from .openrouter import OpenRouterProvider
 from .reka import RekaProvider
 from .together import TogetherProvider
+
+__all__ = [
+    "AI21Provider",
+    "AlephAlphaProvider",
+    "AnthropicProvider",
+    "BedrockAnthropicProvider",
+    "CohereProvider",
+    "GoogleProvider",
+    "GoogleGenAIProvider",
+    "HuggingfaceHubProvider",
+    "OpenAIProvider",
+    "MistralProvider",
+    "OllamaProvider",
+    "DeepSeekProvider",
+    "GroqProvider",
+    "RekaProvider",
+    "TogetherProvider",
+    "OpenRouterProvider",
+]
